split model tests into smaller focused files
